+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 
+## [1.2.29] - Unreleased
+- Add TextEditor::AnyState const value.
+- Add TextEditor::add_key_binding() and remove_key_binding().
+
 ## [1.2.28] - 2022-01-22
 - Expose TreeItem user_data.
 - Documentation fix by @AshfordN.